# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Monthly payout calculation for couriers with tiered commission.",
        "title": "Payout Calculation Service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Couriers always get only their own orders, the courier_id filter is ignored for them.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "List orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courier id",
                        "name": "courier_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pending | delivered | cancelled | returned",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339",
                        "name": "delivered_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339",
                        "name": "delivered_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339",
                        "name": "created_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339",
                        "name": "created_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "decimal",
                        "name": "amount_min",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "decimal",
                        "name": "amount_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "amount | delivered_at | created_at",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc | desc",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "page size, max 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.OrderListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "order",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "courier not found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.OrderResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Delete order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets delivered_at when status becomes delivered. If the order's month already has a payout, an adjustment is written in the same transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Change order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "new status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.updateOrderStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ruziba3vich_payout-calculation-service_internal_domain_order.Status": {
            "type": "string",
            "enum": [
                "pending",
                "delivered",
                "cancelled",
                "returned"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusDelivered",
                "StatusCancelled",
                "StatusReturned"
            ]
        },
...
        "internal_interfaces_http_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_interfaces_http_handler.OrderListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_interfaces_http_handler.OrderResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_interfaces_http_handler.OrderResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "courier_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_ruziba3vich_payout-calculation-service_internal_domain_order.Status"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_interfaces_http_handler.createOrderRequest": {
            "type": "object",
            "required": [
                "amount",
                "courier_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "courier_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_interfaces_http_handler.updateOrderStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "enum": [
                        "pending",
                        "delivered",
                        "cancelled",
                        "returned"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_ruziba3vich_payout-calculation-service_internal_domain_order.Status"
                        }
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by the token from /auth/admin/login or /auth/courier/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}